push next alert on read status
