+++ app/src/main/java/info/nightscout/androidaps/queue/commands/CommandReadStatus.java
@@ -1,8 +1,8 @@
 package info.nightscout.androidaps.queue.commands;
 
-import info.nightscout.androidaps.MainApp;
 import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
 import info.nightscout.androidaps.queue.Callback;
+import info.nightscout.utils.LocalAlertUtils;
 
 /**
  * Created by mike on 09.11.2017.
@@ -20,6 +20,7 @@ public class CommandReadStatus extends Command {
     @Override
     public void execute() {
         ConfigBuilderPlugin.getActivePump().getPumpStatus();
+        LocalAlertUtils.notifyPumpStatusRead();
         if (callback != null)
             callback.result(null).run();
     }
