+++ app/src/main/java/info/nightscout/androidaps/receivers/KeepAliveReceiver.java
@@ -16,34 +16,16 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Date;
 
-import info.nightscout.androidaps.Config;
 import info.nightscout.androidaps.Constants;
 import info.nightscout.androidaps.MainApp;
-import info.nightscout.androidaps.R;
-import info.nightscout.androidaps.db.BgReading;
-import info.nightscout.androidaps.db.DatabaseHelper;
 import info.nightscout.androidaps.interfaces.PumpInterface;
 import info.nightscout.androidaps.data.Profile;
 import info.nightscout.androidaps.plugins.ConfigBuilder.ConfigBuilderPlugin;
-import info.nightscout.androidaps.plugins.Overview.notifications.Notification;
-import info.nightscout.androidaps.plugins.Overview.events.EventNewNotification;
-import info.nightscout.utils.SP;
+import info.nightscout.utils.LocalAlertUtils;
 
 public class KeepAliveReceiver extends BroadcastReceiver {
     private static Logger log = LoggerFactory.getLogger(KeepAliveReceiver.class);
     public static final long STATUS_UPDATE_FREQUENCY = 15 * 60 * 1000L;
-
-    // TODO consider moving this into an Alarms plugin that works offline and can be configured
-    // (e.g. override silent mode at night only)
-
-    private static int missedReadingsThreshold() {
-        return SP.getInt(MainApp.sResources.getString(R.string.key_missed_bg_readings_threshold), 30) * 60 * 1000;
-    }
-
-    private static int pumpUnreachableThreshold() {
-        return SP.getInt(MainApp.sResources.getString(R.string.key_pump_unreachable_threshold), 30) * 60 * 1000;
-    }
-
 
     @Override
     public void onReceive(Context context, Intent rIntent) {
@@ -51,24 +33,12 @@ public class KeepAliveReceiver extends BroadcastReceiver {
         PowerManager.WakeLock wl = pm.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "");
         wl.acquire();
 
-        shortenSnoozeInterval();
-        checkBg();
+        LocalAlertUtils.shortenSnoozeInterval();
+        LocalAlertUtils.checkStaleBGAlert();
         checkPump();
 
         log.debug("KeepAlive received");
         wl.release();
-    }
-
-    private void checkBg() {
-        BgReading bgReading = DatabaseHelper.lastBg();
-        if (SP.getBoolean(MainApp.sResources.getString(R.string.key_enable_missed_bg_readings_alert), false)
-                && bgReading != null && bgReading.date + missedReadingsThreshold() < System.currentTimeMillis()
-                && SP.getLong("nextMissedReadingsAlarm", 0l) < System.currentTimeMillis()) {
-            Notification n = new Notification(Notification.BG_READINGS_MISSED, MainApp.sResources.getString(R.string.missed_bg_readings), Notification.URGENT);
-            n.soundId = R.raw.alarm;
-            SP.putLong("nextMissedReadingsAlarm", System.currentTimeMillis() + missedReadingsThreshold());
-            MainApp.bus().post(new EventNewNotification(n));
-        }
     }
 
     private void checkPump() {
@@ -79,16 +49,7 @@ public class KeepAliveReceiver extends BroadcastReceiver {
             boolean isStatusOutdated = lastConnection.getTime() + STATUS_UPDATE_FREQUENCY < System.currentTimeMillis();
             boolean isBasalOutdated = Math.abs(profile.getBasal() - pump.getBaseBasalRate()) > pump.getPumpDescription().basalStep;
 
-            boolean alarmTimeoutExpired = lastConnection.getTime() + pumpUnreachableThreshold() < System.currentTimeMillis();
-            boolean nextAlarmOccurrenceReached = SP.getLong("nextPumpDisconnectedAlarm", 0l) < System.currentTimeMillis();
-
-            if (Config.APS && SP.getBoolean(MainApp.sResources.getString(R.string.key_enable_pump_unreachable_alert), true)
-                    && isStatusOutdated && alarmTimeoutExpired && nextAlarmOccurrenceReached && !ConfigBuilderPlugin.getActiveLoop().isDisconnected()) {
-                Notification n = new Notification(Notification.PUMP_UNREACHABLE, MainApp.sResources.getString(R.string.pump_unreachable), Notification.URGENT);
-                n.soundId = R.raw.alarm;
-                SP.putLong("nextPumpDisconnectedAlarm", System.currentTimeMillis() + pumpUnreachableThreshold());
-                MainApp.bus().post(new EventNewNotification(n));
-            }
+            LocalAlertUtils.checkPumpUnreachableAlarm(lastConnection, isStatusOutdated);
 
             if (!pump.isThisProfileSet(profile)) {
                 MainApp.getConfigBuilder().getCommandQueue().setProfile(profile, null);
@@ -103,8 +64,8 @@ public class KeepAliveReceiver extends BroadcastReceiver {
     //called by MainApp at first app start
     public void setAlarm(Context context) {
 
-        shortenSnoozeInterval();
-        presnoozeAlarms();
+        LocalAlertUtils.shortenSnoozeInterval();
+        LocalAlertUtils.presnoozeAlarms();
 
         AlarmManager am = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
         Intent i = new Intent(context, KeepAliveReceiver.class);
@@ -117,33 +78,4 @@ public class KeepAliveReceiver extends BroadcastReceiver {
         am.setInexactRepeating(AlarmManager.RTC_WAKEUP, System.currentTimeMillis(), Constants.keepAliveMsecs, pi);
     }
 
-    /*Presnoozes the alarms with 5 minutes if no snooze exists.
-     * Call only at startup!
-     */
-    public void presnoozeAlarms() {
-        if (SP.getLong("nextMissedReadingsAlarm", 0l) < System.currentTimeMillis()) {
-            SP.putLong("nextMissedReadingsAlarm", System.currentTimeMillis() + 5 * 60 * 1000);
-        }
-        if (SP.getLong("nextPumpDisconnectedAlarm", 0l) < System.currentTimeMillis()) {
-            SP.putLong("nextPumpDisconnectedAlarm", System.currentTimeMillis() + 5 * 60 * 1000);
-        }
-    }
-
-    public void cancelAlarm(Context context) {
-        Intent intent = new Intent(context, KeepAliveReceiver.class);
-        PendingIntent sender = PendingIntent.getBroadcast(context, 0, intent, 0);
-        AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
-        alarmManager.cancel(sender);
-    }
-
-    static void shortenSnoozeInterval() {
-        //shortens alarm times in case of setting changes or future data
-        long nextMissedReadingsAlarm = SP.getLong("nextMissedReadingsAlarm", 0L);
-        nextMissedReadingsAlarm = Math.min(System.currentTimeMillis() + missedReadingsThreshold(), nextMissedReadingsAlarm);
-        SP.putLong("nextMissedReadingsAlarm", nextMissedReadingsAlarm);
-
-        long nextPumpDisconnectedAlarm = SP.getLong("nextPumpDisconnectedAlarm", 0L);
-        nextPumpDisconnectedAlarm = Math.min(System.currentTimeMillis() + pumpUnreachableThreshold(), nextPumpDisconnectedAlarm);
-        SP.putLong("nextPumpDisconnectedAlarm", nextPumpDisconnectedAlarm);
-    }
 }
